LR-1016: Added react-datepicker.Small separations in time-line components
