+++ src/@components/time-line/time-line.tsx
@@ -5,15 +5,14 @@ import get from 'lodash/get';
 import minBy from 'lodash/minBy';
 import moment from 'moment';
 
-import grey from '@material-ui/core/colors/grey';
-import orange from '@material-ui/core/colors/orange';
 import Paper from '@material-ui/core/Paper';
 import { Theme, useTheme } from '@material-ui/core/styles';
 
 import CurrentTimeIndicator from './current-time-indicator/current-time-indicator';
 import EditWork from './edit-work';
-import HoverInfo from './hover-info';
 import { useStyles } from './styles';
+import { SvgTimeScale } from './svg-time-scale/svg-time-scale';
+import { UserTasks } from './user-tasks/user-tasks';
 
 import { IEvent } from '@types';
 
@@ -29,11 +28,7 @@ export interface IDailyRoutineProps {
 export const Y_HEIGHT_BIG = 65;
 const Y_HEIGHT_LITTLE = 8;
 const X_OFFSET = 24;
-const LABEL_HEIGHT = 12;
-
-const heightTimer: any = null;
 let updateInterval: any = null;
-let leaveTimer: any = null;
 
 const timelineStub = () => null;
 
@@ -62,8 +57,6 @@ export const TimeLineTsx: React.FC<IDailyRoutineProps> = ({
   const [startAt, setStartAt] = useState(getStartAt);
   const [finishAt, setFinishAt] = useState(getFinishAt);
   const [height] = useState(fullSize ? Y_HEIGHT_BIG : Y_HEIGHT_LITTLE);
-  const [hoveredEl, setHoveredEl] = useState<any>(null);
-  const [hoveredEvent, setHoveredEvent] = useState<IEvent>();
   const [editedEvent, setEditedEvent] = useState<IEvent>();
 
   const handleEditEventClose = useCallback(
@@ -80,60 +73,6 @@ export const TimeLineTsx: React.FC<IDailyRoutineProps> = ({
       onTimelineClick();
     }
   }, [fullSize, onTimelineClick]);
-
-  const handlePopoverClose = useCallback(() => {
-    leaveTimer = setTimeout(() => {
-      setHoveredEl(null);
-      setHoveredEvent(undefined);
-    }, 500);
-
-    return () => {
-      clearTimeout(leaveTimer);
-    };
-  }, [setHoveredEl, setHoveredEvent]);
-
-  const cleanLeaveTimer = useCallback(() => {
-    if (leaveTimer) {
-      clearTimeout(leaveTimer);
-    }
-  }, []);
-
-  const handleHover = useCallback(
-    (e: React.SyntheticEvent) => {
-      const eventId = parseInt(get(e, ['target', 'dataset', 'id'], 0), 0);
-      const newHoveredEvent = events.find(el => get(el, ['userWork', 'id']) === eventId);
-      cleanLeaveTimer();
-      setHoveredEl(e.currentTarget);
-      if (!editedEvent && newHoveredEvent) {
-        setHoveredEvent(newHoveredEvent);
-      } else {
-        setHoveredEvent(undefined);
-      }
-    },
-    [cleanLeaveTimer, editedEvent, events]
-  );
-
-  const getStyle = useCallback((taskInfo: IEvent) => {
-    if (taskInfo.isActive) {
-      return {
-        backgroundColor: '#FFF0B5',
-        borderBottomWidth: 1,
-        borderColor: '#FFB200',
-        borderLeftWidth: 2,
-        borderRadius: 4,
-        borderRightWidth: 2,
-        borderStyle: 'solid',
-        borderTopWidth: 1,
-      };
-    }
-    return {
-      backgroundColor: '#D5D5D5',
-      borderColor: grey[400],
-      borderRadius: 4,
-      borderStyle: 'solid',
-      borderWidth: 1,
-    };
-  }, []);
 
   const svgWidth = useMemo(() => {
     return width - 2 * X_OFFSET;
@@ -159,57 +98,12 @@ export const TimeLineTsx: React.FC<IDailyRoutineProps> = ({
     [finishAt, getHours, startAt, svgWidth]
   );
 
-  const getWidth = useCallback(
-    (el: IEvent) => {
-      return getPosition(el.userWork.finishAt) - getPosition(el.userWork.startAt);
-    },
-    [getPosition]
-  );
-
-  const handleEventClick = useCallback(
-    (e: React.SyntheticEvent) => {
-      const eventId = parseInt(get(e, ['target', 'dataset', 'id'], 0), 0);
-      if (height === Y_HEIGHT_BIG) {
-        e.stopPropagation();
-        if (eventId === get(editedEvent, ['userWork', 'id'])) {
-          setEditedEvent(undefined);
-        } else {
-          const newEditEvent = events.find(el => get(el, ['userWork', 'id']) === eventId);
-          if (newEditEvent) {
-            setEditedEvent(newEditEvent);
-            setHoveredEvent(undefined);
-          }
-        }
-      }
-    },
-    [editedEvent, events, height, setEditedEvent, setHoveredEvent]
-  );
-
-  const getLines = useCallback(() => {
-    // Узнаем разницу между текущим моментом и моментом начала задачи
-    const parts = (finishAt - startAt) * 4;
-    if (parts <= 0) {
-      return [];
-    }
-    const step = svgWidth / parts;
-    const arr = new Array(parts).fill(0).map((_, i) => ({
-      isHour: !(i % 4),
-      label: !(i % 4) && `${Math.ceil(i / 4 + startAt)}`,
-      x: step * i,
-    }));
-    arr.push({ x: svgWidth, isHour: true, label: `${finishAt}` });
-    return arr;
-  }, [finishAt, startAt, svgWidth]);
-
   useEffect(() => {
     updateInterval = setInterval(() => {
       setStartAt(getStartAt());
       setFinishAt(getFinishAt());
     }, 600000);
     return () => {
-      if (heightTimer) {
-        clearTimeout(heightTimer);
-      }
       if (updateInterval) {
         clearInterval(updateInterval);
       }
@@ -219,23 +113,13 @@ export const TimeLineTsx: React.FC<IDailyRoutineProps> = ({
   const classes = useStyles();
   const theme: Theme = useTheme();
 
-  const preparedEvents = useMemo(() => {
-    return events.filter((el: IEvent) => {
-      return (
-        !el.userWork.finishAt ||
-        (el.userWork.finishAt.day() === moment().day() && getHours(el.userWork.finishAt) > startAt)
-      );
-    });
-  }, [events, getHours, startAt]);
-
-  const isExpended = useMemo(() => height === Y_HEIGHT_BIG, [height]);
-
   const formInitialValues = useMemo(() => {
     return {
       ...get(editedEvent, ['userWork']),
       projectId: get(editedEvent, ['userWork', 'projectId'], get(editedEvent, ['task', 'projectId'])),
     };
   }, [editedEvent]);
+
   return (
     <Popover
       style={{ zIndex: 1400 }}
@@ -255,75 +139,42 @@ export const TimeLineTsx: React.FC<IDailyRoutineProps> = ({
     >
       <div className={classes.rootWrap} onClick={handleClick}>
         <CurrentTimeIndicator fullSize={fullSize} left={getPosition(moment())} />
-
         <div
           ref={getRef}
           className={classes.root}
           style={{
             height,
-            zIndex: isExpended ? 1200 : 0,
+            zIndex: fullSize ? 1200 : 0,
           }}
         >
           <div
             className={classes.filled}
             style={{
-              boxShadow: isExpended ? theme.shadows[1] : 'none',
-              flexBasis: isExpended ? '76%' : '100%',
+              boxShadow: fullSize ? theme.shadows[1] : 'none',
+              flexBasis: fullSize ? '76%' : '100%',
             }}
           >
-            {preparedEvents.map((taskInfo, i) => {
-              return (
-                <Popover
-                  key={taskInfo.userWork.id}
-                  preferPlace="below"
-                  tipSize={0.01}
-                  className={classes.popover}
-                  isOpen={get(hoveredEvent, 'userWork.id') === taskInfo.userWork.id && !!hoveredEvent}
-                  target={hoveredEl as any}
-                  onOuterAction={handlePopoverClose}
-                  body={<HoverInfo onOver={handleHover} onLeave={handlePopoverClose} hoveredEvent={taskInfo} />}
-                >
-                  <div
-                    aria-owns={`popover-body-${taskInfo.userWork.id}`}
-                    data-id={taskInfo.userWork.id}
-                    className={classes.block}
-                    style={{
-                      ...getStyle(taskInfo),
-                      left: getPosition(taskInfo.userWork.startAt),
-                      width: getWidth(taskInfo),
-                    }}
-                    onClick={handleEventClick}
-                    onMouseOver={handleHover}
-                    onMouseLeave={handlePopoverClose}
-                  />
-                </Popover>
-              );
-            })}
+            <UserTasks
+              editedEvent={editedEvent}
+              events={events}
+              height={height}
+              getPosition={getPosition}
+              setEditedEvent={setEditedEvent}
+              startAt={startAt}
+              getHours={getHours}
+              Y_HEIGHT_BIG={Y_HEIGHT_BIG}
+            />
           </div>
-          <svg height={fullSize ? 80 : height} width={width} className={classes.svg}>
-            {getLines().map(({ x, isHour, label }) => (
-              <React.Fragment key={x}>
-                {label && height === Y_HEIGHT_BIG && (
-                  <text x={x + X_OFFSET} y={10} className={classes.text}>
-                    <tspan x={x + X_OFFSET} textAnchor="middle">
-                      {label}
-                      :00
-                    </tspan>
-                  </text>
-                )}
-                {isExpended && (
-                  <line
-                    // stroke="#FAB203"
-                    stroke={orange[300]}
-                    x1={x + X_OFFSET}
-                    y1={LABEL_HEIGHT + 6}
-                    x2={x + X_OFFSET}
-                    y2={LABEL_HEIGHT - 4}
-                  />
-                )}
-              </React.Fragment>
-            ))}
-          </svg>
+          <SvgTimeScale
+            finishAt={finishAt}
+            fullSize={fullSize}
+            height={height}
+            startAt={startAt}
+            svgWidth={svgWidth}
+            width={width}
+            X_OFFSET={X_OFFSET}
+            Y_HEIGHT_BIG={Y_HEIGHT_BIG}
+          />
         </div>
       </div>
     </Popover>
